Allow apps to control their scale. This change fixes the low video frame rate issue by allowing the application image be scaled by the graphics card. There are now two components that determine the size of the application:
the preferredWidth and preferredHeight in pixels
the pixel scale (a scale factor which defaults to 1.0f)

These are set in the WFS properties file.

In the video application, the video is scaled from its native resolution (e.g., 320x240) to a larger scale (e.g., 1024x768) and then this image is scaled by the pixelScale into physical units. By default the scale is 256 pixels to 1 physical unit.

Best performance is achieved when pixelScale is used in preference to preferredWidth and preferredHeight
Best rendering quality is achieved when preferredWidth and preferredHeight are used to scale the image.

Pick one!