Fix bug #369: PDF viewer intermittently fails to load document
Correctly handle a request denied message from the PDF GLO when a request from another client is in progress. This happens when many users join the world at the same time and attempt to get the state of the same PDF document. The solution is to implement a request retry mechanism, as used in the panorama video application. This means that if multiple users press next/previous page, all requests will be honored.

Change debug level of certain messages to verify that this fix is working as expected. The log level will be reduced once the fix has been verified.