Fix bug #420: HUD buttons become visible without their control panels
When a collaborative app receives a message which affects the state of the control panel buttons, the app wasn't checking to see if the button was currently visible. So, for example, when the PDF viewer entered slide show mode, the PDF viewer would show the pause button regardless of the state of the HUD. It now only updates the button state if the HUD is visible or when the HUD is opened.

This fix has already been applied to the video app.